Переместила пакеты utils, model, client в src.main.java
