+++ src/test/java/project/ordertests/CreateOrderTest.java
@@ -5,7 +5,7 @@ import io.restassured.RestAssured;
 import io.restassured.response.Response;
 import org.junit.Before;
 import org.junit.Test;
-import project.helpers.OrderNoColor;
+import model.OrderNoColor;
 
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.notNullValue;
